Extract seed routes out of agent-channel.ts (Wave D batch 4)
agent-channel.ts 2877 -> 2771 lines. Moves SEED/SEED_REGENERATE into a
new agent-channel-seed.ts as registerAgentChannelSeedRoutes(app, deps),
same pattern as batches 1-3.

Neither route touches channelState/stopReason/the rate-limit Maps.
store and gamesStore are passed by reference (| undefined, matching
agent-channel-brief.ts's store pattern) since both are shared with
other routes still in the parent file; onRegenerateSeed is exclusive
to this cluster. RegenerateSeedRequestSchema/REGENERATE_SEED_REFUSALS
move with the cluster (their only callers).
